apps/pyproject.toml
[project]
name = "agent-apps"
version = "0.1.0"
description = "Demo applications for the tool-calling agent"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.45.0",
    "python-dotenv>=1.2.2",
    "agent-core",
    "react-agent",
    "planning-agent",
]

[tool.uv.sources]
agent-core = { path = "../lib", editable = true }
react-agent = { path = "../lib/react_agent", editable = true }
planning-agent = { path = "../lib/planning_agent", editable = true }

[tool.ruff]
line-length = 100
